Separated components of login page into separate files.
